Add option where clicks apply filter instantly
This kept coming up with users in over the shoulder testing
where they expected clicks to filters to be applied instantly.
To remedy this have added a checkbox which can turn on instant
filters or turn them off. Saved to localstorage to try and make
things nicer for the user by by default is on.