Add file index repo job type allowing for index filepaths
Request from potential customer asking for the ability to index
arbitary filepaths. To do so we create a new repojob which performs
no updates and just does full filepath indexing. The reason for the
new job to enqueue this is to make run on a slower schedule because
the operation is costly as it cannot perform deltas.